Simplifying component structure in prep for further divergence.
